requirements.txt
numpy>=1.12.1
opencv-python>=3.2.0.7
paho-mqtt>=1.5.1
Pillow>=6.2.0